TST GH26807 move pytest fixtures to conftest
